+++ src/neo4j/exceptions.py
@@ -51,6 +51,7 @@ Driver API Errors
     + RoutingServiceUnavailable
     + WriteServiceUnavailable
     + ReadServiceUnavailable
+    + UnsupportedServerProduct
     + IncompleteCommit
   + ConfigurationError
     + AuthConfigurationError
@@ -1146,6 +1147,17 @@ class ReadServiceUnavailable(ServiceUnavailable):
     """Raised when no read service is available."""
 
 
+# DriverError > ServiceUnavailable > UnsupportedServerProduct
+class UnsupportedServerProduct(ServiceUnavailable):
+    """
+    Raised when an unsupported server product is detected.
+
+    .. versionchanged:: 6.0
+        This exception is now a subclass of :class:`ServiceUnavailable`.
+        Before it was a subclass of :class:`Exception`.
+    """
+
+
 # DriverError > ServiceUnavailable > IncompleteCommit
 class IncompleteCommit(ServiceUnavailable):
     """
@@ -1184,7 +1196,3 @@ class AuthConfigurationError(ConfigurationError):
 # DriverError > ConfigurationError > CertificateConfigurationError
 class CertificateConfigurationError(ConfigurationError):
     """Raised when there is an error with the certificate configuration."""
-
-
-class UnsupportedServerProduct(Exception):
-    """Raised when an unsupported server product is detected."""